feat: change idadeMediaAlunos type from int to double in GeralDashboardResponse

+++ src/main/java/com/fighthub/dto/dashboard/GeralDashboardResponse.java
@@ -6,7 +6,7 @@ public record GeralDashboardResponse(
         int totalAlunosInativos,
         int novosAlunosUltimos30Dias,
         int alunosQueSairamUltimos30Dias,
-        int idadeMediaAlunos
+        double idadeMediaAlunos
 
 ) {
 }
